// frontend/src/app/openapi/models/Graph.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type Graph = {
  name: string;
  description: (string | null);
  is_favorite?: boolean;
  uuid: string;
  updated: string;
  created: string;
  last_seen: string;
};